+++ src/components/about_page_comoponents/ShortHero.jsx
